+++ src/lib/components/grades/gradeCard.svelte
@@ -5,6 +5,8 @@
 	export let semesterName: string;
 	import AppCard from '$shared/AppCard.svelte';
 	import * as allIonicIcons from 'ionicons/icons';
+	import Course from '../../../routes/courses/[courseId]/courses.svelte';
+	import {navController} from 'ionic-svelte';
 
 
 	let childrenOpen: boolean[] = [];
@@ -12,6 +14,10 @@
     function toggleChildren(index: number) {
         childrenOpen[index] = !childrenOpen[index];
     }
+
+	function navigateToCourse(course: { course: any; }) {
+		navController.push(Course, {id: course.course});
+	}
 
 </script>
 <span class="scroll" id={semesterId}></span>
@@ -35,11 +41,13 @@
 </div>
 
 	{#each filteredSubjects as course,index}
-	<AppCard id={index} href={course.childCourses && course.childCourses.length > 0 ? undefined : `/courses/${encodeURIComponent(course.course)}`} padding>
+	<AppCard id={index} onClick={() => navigateToCourse(course)} padding>
 		<!-- Card content for course -->
 		<!-- Checking if course has children courses or not, so we can render the href links accordingly. Rest of the content stays the same -->
 		{#if course.childCourses && course.childCourses.length > 0}
-		<ion-item href={ `/courses/${encodeURIComponent(course.course)}`} lines="none" class="ion-no-padding">
+		<!-- svelte-ignore a11y-no-static-element-interactions -->
+		<!-- svelte-ignore a11y-click-events-have-key-events -->
+		<ion-item on:click={navigateToCourse} lines="none" class="ion-no-padding">
 			<div class="containerFlex">
 
 				<div class="titlesFlex">
